DiscountCodeDraft to interface. #942
