+++ sphere-models/src/it/java/io/sphere/sdk/carts/CartFixtures.java
@@ -14,7 +14,7 @@ import io.sphere.sdk.carts.commands.updateactions.*;
 import io.sphere.sdk.client.BlockingSphereClient;
 import io.sphere.sdk.customers.Customer;
 import io.sphere.sdk.discountcodes.DiscountCode;
-import io.sphere.sdk.discountcodes.DiscountCodeDraft;
+import io.sphere.sdk.discountcodes.DiscountCodeDraftDsl;
 import io.sphere.sdk.discountcodes.commands.DiscountCodeCreateCommand;
 import io.sphere.sdk.discountcodes.commands.DiscountCodeDeleteCommand;
 import io.sphere.sdk.models.Address;
@@ -158,7 +158,7 @@ public class CartFixtures {
                     .validUntil(null)
                     .build();
             final CartDiscount cartDiscount = client.executeBlocking(CartDiscountCreateCommand.of(draft));
-            final DiscountCode discountCode = client.executeBlocking(DiscountCodeCreateCommand.of(DiscountCodeDraft.of(randomKey(), cartDiscount)));
+            final DiscountCode discountCode = client.executeBlocking(DiscountCodeCreateCommand.of(DiscountCodeDraftDsl.of(randomKey(), cartDiscount)));
             final Cart updatedCart = user.apply(cart, discountCode);
             client.executeBlocking(CartDeleteCommand.of(updatedCart));
             client.executeBlocking(DiscountCodeDeleteCommand.of(discountCode));
