+++ sphere-models/src/main/java/io/sphere/sdk/discountcodes/DiscountCodeDraft.java
@@ -1,9 +1,8 @@
 package io.sphere.sdk.discountcodes;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import io.sphere.sdk.cartdiscounts.CartDiscount;
-import io.sphere.sdk.cartdiscounts.CartDiscountPredicate;
-import io.sphere.sdk.models.Base;
 import io.sphere.sdk.models.LocalizedString;
 import io.sphere.sdk.models.Reference;
 import io.sphere.sdk.models.Referenceable;
@@ -14,117 +13,37 @@ import java.util.List;
 
 /**
  * @see DiscountCodeDraftBuilder
+ * @see DiscountCodeDraftDsl
  */
-public class DiscountCodeDraft extends Base {
+@JsonDeserialize(as = DiscountCodeDraftDsl.class)
+public interface DiscountCodeDraft {
     @Nullable
-    private final LocalizedString name;
-    @Nullable
-    private final LocalizedString description;
-    private final String code;
-    private final List<Reference<CartDiscount>> cartDiscounts;
-    @Nullable
-    private final String cartPredicate;
-    private final Boolean isActive;
-    @Nullable
-    private final Long maxApplications;
-    @Nullable
-    private final Long maxApplicationsPerCustomer;
-
-    DiscountCodeDraft(final List<Reference<CartDiscount>> cartDiscounts, @Nullable final LocalizedString name, @Nullable final LocalizedString description, final String code, @Nullable final String cartPredicate, final Boolean isActive, @Nullable final Long maxApplications, @Nullable final Long maxApplicationsPerCustomer) {
-        this.cartDiscounts = cartDiscounts;
-        this.name = name;
-        this.description = description;
-        this.code = code;
-        this.cartPredicate = cartPredicate;
-        this.isActive = isActive;
-        this.maxApplications = maxApplications;
-        this.maxApplicationsPerCustomer = maxApplicationsPerCustomer;
-    }
-
-    public static DiscountCodeDraft of(final String code, final Referenceable<CartDiscount> cartDiscount) {
-        return of(code, Collections.singletonList(cartDiscount.toReference()));
-    }
-
-    public static DiscountCodeDraft of(final String code, final List<Reference<CartDiscount>> cartDiscounts) {
-        return DiscountCodeDraftBuilder.of(code, cartDiscounts).build();
-    }
-
-    public DiscountCodeDraft withName(@Nullable final LocalizedString name) {
-        return newBuilder().name(name).build();
-    }
+    LocalizedString getName();
 
     @Nullable
-    public LocalizedString getName() {
-        return name;
-    }
+    LocalizedString getDescription();
 
-    public DiscountCodeDraft withDescription(@Nullable final LocalizedString description) {
-        return newBuilder().description(description).build();
-    }
+    String getCode();
+
+    List<Reference<CartDiscount>> getCartDiscounts();
 
     @Nullable
-    public LocalizedString getDescription() {
-        return description;
-    }
-
-    public DiscountCodeDraft withCode(final String code) {
-        return newBuilder().code(code).build();
-    }
-
-    public String getCode() {
-        return code;
-    }
-
-    public DiscountCodeDraft withCartDiscounts(final Referenceable<CartDiscount> cartDiscount) {
-        return withCartDiscounts(Collections.singletonList(cartDiscount.toReference()));
-    }
-
-    public DiscountCodeDraft withCartDiscounts(final List<Reference<CartDiscount>> cartDiscounts) {
-        return newBuilder().cartDiscounts(cartDiscounts).build();
-    }
-
-    public List<Reference<CartDiscount>> getCartDiscounts() {
-        return cartDiscounts;
-    }
-
-
-    public DiscountCodeDraft withCartPredicate(@Nullable final CartDiscountPredicate cartPredicate) {
-        return newBuilder().cartPredicate(cartPredicate).build();
-    }
-
-    @Nullable
-    public String getCartPredicate() {
-        return cartPredicate;
-    }
-
-    public DiscountCodeDraft withIsActive(final Boolean isActive) {
-        return newBuilder().isActive(isActive).build();
-    }
+    String getCartPredicate();
 
     @JsonProperty("isActive")
-    public Boolean isActive() {
-        return isActive;
-    }
-
-    public DiscountCodeDraft withMaxApplications(@Nullable final Long maxApplications) {
-        return newBuilder().maxApplications(maxApplications).build();
-    }
+    Boolean isActive();
 
     @Nullable
-    public Long getMaxApplications() {
-        return maxApplications;
-    }
-
-    public DiscountCodeDraft withMaxApplicationsPerCustomer(@Nullable final Long maxApplicationsPerCustomer) {
-        return newBuilder().maxApplicationsPerCustomer(maxApplicationsPerCustomer).build();
-    }
+    Long getMaxApplications();
 
     @Nullable
-    public Long getMaxApplicationsPerCustomer() {
-        return maxApplicationsPerCustomer;
+    Long getMaxApplicationsPerCustomer();
+
+    static DiscountCodeDraftDsl of(final String code, final Referenceable<CartDiscount> cartDiscount) {
+        return DiscountCodeDraftDsl.of(code, cartDiscount);
     }
 
-    private DiscountCodeDraftBuilder newBuilder() {
-        return DiscountCodeDraftBuilder.of(this);
+    static DiscountCodeDraftDsl of(final String code, final List<Reference<CartDiscount>> cartDiscounts) {
+        return DiscountCodeDraftDsl.of(code, cartDiscounts);
     }
 }
